sistema_login/requirements.txt
flask
mysql-connector-python
werkzeug
flask-cors
